% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{AddDayNightInfo}
\alias{AddDayNightInfo}
\title{Finds if an opbservation falls into the day time or not}
\usage{
AddDayNightInfo(fileInfoDF)
}
\arguments{
\item{trainTestDT}{Data Table with image infromation}
}
\description{
Finds if an opbservation falls into the day time or not
}

