% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machineLearningRelated.R
\name{assignTrainAndTestSet}
\alias{assignTrainAndTestSet}
\title{Divide the training and test set based on day number}
\usage{
assignTrainAndTestSet(direcoryPath)
}
\arguments{
\item{fullFilename}{String}

\item{pattern}{String encoding the filename pattern}
}
\description{
Divide the training and test set based on day number
}

