% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ImageStatistics.R
\name{CalculateVarGradient}
\alias{CalculateVarGradient}
\title{Calculate variance of gradient}
\usage{
CalculateVarGradient(filename, axes = "x")
}
\description{
Calculate variance of gradient
}

