% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{Dehaze}
\alias{Dehaze}
\title{Obtain the dehazed image}
\usage{
Dehaze(image, omega, winSize, lambda)
}
\arguments{
\item{image}{Image}

\item{omega}{Constant for aerial perspective}

\item{winSize}{Should probably be renamed to patch}

\item{lambda}{Regularization parameter for soft matting}
}
\description{
Obtain the dehazed image
}

