% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEdges.R
\name{DetectMeanEdges}
\alias{DetectMeanEdges}
\title{Return mean number of edges}
\usage{
DetectMeanEdges(im, sigma = 1)
}
\arguments{
\item{im}{Image}

\item{sigma}{Sigma}
}
\description{
Return mean number of edges
}
