% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageStatistics.R
\name{FileNameParser}
\alias{FileNameParser}
\title{Extracts name and date time from filename}
\usage{
FileNameParser(fullFilename, pattern = "na*me_yyyymmdd_hhmm.jpg")
}
\arguments{
\item{fullFilename}{String}

\item{pattern}{String encoding the filename pattern}
}
\description{
Extracts name and date time from filename
}

