% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{FilterDayLightHours}
\alias{FilterDayLightHours}
\title{Filter the lighthours given the information of the station (longitude, latitude) and day}
\usage{
FilterDayLightHours(mergedData, originalFileInfoDT)
}
\arguments{
\item{mergedData}{data of station of interest and day of interest}

\item{originalFileInfoDT}{data table with image files information}
}
\description{
Filter the lighthours given the information of the station (longitude, latitude) and day
}

