% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portingMatlabCode.R
\name{GetAtmosphere}
\alias{GetAtmosphere}
\title{Obtains atmosphere}
\usage{
GetAtmosphere(image, darkChannel)
}
\arguments{
\item{image}{The image object}

\item{darkChannel}{Image dark channel}
}
\description{
Obtains atmosphere
}

