% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{GetDarkChannel}
\alias{GetDarkChannel}
\title{Obtains the dark channel}
\usage{
GetDarkChannel(image, winSize)
}
\arguments{
\item{image}{The image object}

\item{winSize}{Should probably be renamed}
}
\description{
Obtains the dark channel
}

