% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{GetFractalDim}
\alias{GetFractalDim}
\title{Fractal dimension of numeric vector or cimg}
\usage{
GetFractalDim(x)
}
\arguments{
\item{x}{Numeric vector, matrix or of class cimg}
}
\description{
Fractal dimension of numeric vector or cimg
}
\note{
Images (e.g objects of class cimg) are first transformed to grayscale
 and then considered as a matrix
}
