% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{GetHorizAvgTrans}
\alias{GetHorizAvgTrans}
\title{Obtains the dark horizontal average transmission}
\usage{
GetHorizAvgTrans(image, winSize = 15, omega = 0.95, lambda = 0.001)
}
\arguments{
\item{image}{The image object}

\item{winSize}{Should probably be renamed}

\item{omega}{omega parameter}

\item{lambda}{lambda parameter}
}
\value{
vertical transmission profile
}
\description{
Obtains the dark horizontal average transmission
}
