% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{GetMattingLaplacian}
\alias{GetMattingLaplacian}
\title{Obtains Matting Laplacian of the image}
\usage{
GetMattingLaplacian(image)
}
\arguments{
\item{image}{Image}
}
\description{
Obtains Matting Laplacian of the image
}
\references{
\url{http://www.wisdom.weizmann.ac.il/~levina/papers/Matting-Levin-Lischinski-Weiss-CVPR06.pdf}
}

