% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{GetRadiance}
\alias{GetRadiance}
\title{Obtains radiance}
\usage{
GetRadiance(image, transmission = NULL, atmosphere = NULL, omega = 0.95,
  winSize = 15)
}
\arguments{
\item{image}{The image object}

\item{transmission}{The image transmission}

\item{atmosphere}{The image atmosphere}

\item{omega}{Constant for aerial perspective}

\item{winSize}{Should probably be renamed to patch}
}
\description{
Obtains radiance
}
