% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{GetSunTimes}
\alias{GetSunTimes}
\title{Find the sunrise and sunset times given a date and lon lat location}
\usage{
GetSunTimes(data)
}
\arguments{
\item{data}{Data Table with image infromation}
}
\description{
Find the sunrise and sunset times given a date and lon lat location
}

