% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{GetTransmissionEstimate}
\alias{GetTransmissionEstimate}
\title{Obtains transmission}
\usage{
GetTransmissionEstimate(image, atmosphere, omega, winSize)
}
\arguments{
\item{image}{Image}

\item{atmosphere}{The image atmosphere}

\item{omega}{Constant for aerial perspective}

\item{winSize}{Should probably be renamed to patch}
}
\description{
Obtains transmission
}

