% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{HorizontalAverageTransmission}
\alias{HorizontalAverageTransmission}
\title{Obtains horizontal average from transmission}
\usage{
HorizontalAverageTransmission(x)
}
\arguments{
\item{x}{the transmission}
}
\description{
Obtains horizontal average from transmission
}
