% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReturnFeatures.R
\name{ImageFeatures}
\alias{ImageFeatures}
\title{Return image features}
\usage{
ImageFeatures(filePath, ...)
}
\arguments{
\item{filePath}{string}

\item{...}{additional arguments for cropping using imsub}
}
\description{
Return image features
}

