% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{IsDayLightImage}
\alias{IsDayLightImage}
\title{Wich images are between sunrise and sunset?}
\usage{
IsDayLightImage(dateTime, lon, lat)
}
\arguments{
\item{dateTime}{date and time}

\item{lon}{Longitude}

\item{lat}{Latitude}
}
\description{
Wich images are between sunrise and sunset?
}

