% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{MergeDaysWithStationConfig}
\alias{MergeDaysWithStationConfig}
\title{Merge the information of configuration file and days and station of interest}
\usage{
MergeDaysWithStationConfig(dateStation, configDF)
}
\arguments{
\item{dateStation}{Data Table with date of observation and station}

\item{configDF}{Data Table with station information (e.g., lon, lat, id)}
}
\description{
Merge the information of configuration file and days and station of interest
}

