% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PointAnalyser}
\alias{PointAnalyser}
\title{(R,G,B) intensity profiles}
\usage{
PointAnalyser(filename)
}
\arguments{
\item{filename}{The name of the image file.}
}
\description{
With this function one can analyse by clicking on a point in
the selected image the (R,G,B) intensity profiles of the corresponding
image line (in another window)
}

