% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSensorData.R
\name{ReadMORSensorData}
\alias{ReadMORSensorData}
\title{Read MOR data files}
\usage{
ReadMORSensorData(filenames)
}
\arguments{
\item{filenames}{List of filenames}
}
\value{
data.table
}
\description{
Read MOR data files
}

