% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSensorData.R
\name{ReadMORSensorDataTWE}
\alias{ReadMORSensorDataTWE}
\title{Read MOR data files Twente | TEMPORARY TO BE THE ONLY FORMAT SUPPORTED SINCE OFFICIAL FORMAT IN KIS}
\usage{
ReadMORSensorDataTWE(filenames)
}
\arguments{
\item{filenames}{List of filenames}
}
\value{
data.table
}
\description{
Read MOR data files Twente | TEMPORARY TO BE THE ONLY FORMAT SUPPORTED SINCE OFFICIAL FORMAT IN KIS
}

