% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portingMatlabCode.R
\name{RefineTransmissionEstimate}
\alias{RefineTransmissionEstimate}
\title{Refines transmission estimate via matting Laplacian}
\usage{
RefineTransmissionEstimate(image, transmission, lambda)
}
\arguments{
\item{image}{Image}

\item{transmission}{Initial transmission estimate}

\item{lambda}{Regularization parameter for the soft matting}
}
\description{
Refines transmission estimate via matting Laplacian
}

