% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageStatistics.R, R/sunRelated.R
\name{RetriveDaysAndStation}
\alias{RetriveDaysAndStation}
\title{Finds if an opbservation falls into the day time or not}
\usage{
RetriveDaysAndStation(dateStation, configDF)

RetriveDaysAndStation(dateStation, configDF)
}
\arguments{
\item{trainTestDT}{Data Table with image infromation}

\item{trainTestDT}{Data Table with image infromation}
}
\description{
Finds if an opbservation falls into the day time or not

Merge the information of configuration file and days and station of interest
}

