% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSensorData.R
\name{SynchronizeSensorPicture}
\alias{SynchronizeSensorPicture}
\title{Synchronize the sensor reading to picture acquisition time}
\usage{
SynchronizeSensorPicture(sensorDataDT, imageInfoDT)
}
\arguments{
\item{sensorDataDT}{data table with sensor data}

\item{imageInfoDT}{data table with image summary information}
}
\value{
data.table
}
\description{
Synchronize the sensor reading to picture acquisition time
}
