% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{TimeWindowFilter}
\alias{TimeWindowFilter}
\title{Filter the images based on the time window of interest}
\usage{
TimeWindowFilter(originalFileInfoDT, initialTime, finalTime)
}
\arguments{
\item{originalFileInfoDT}{data table with image files information}

\item{initialTime}{initial time in HH:MM}

\item{finalTime}{final time in HH:MM}
}
\description{
Filter the images based on the time window of interest
}

