% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{TransmissionSmoothness}
\alias{TransmissionSmoothness}
\title{Smoothness of transmission}
\usage{
TransmissionSmoothness(transmission)
}
\arguments{
\item{transmission}{The horizontally averaged transmission}
}
\description{
Smoothness of transmission
}

