% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageStatistics.R
\name{UniqueDaysAndStation}
\alias{UniqueDaysAndStation}
\title{Finds if an observation falls into the day time or not}
\usage{
UniqueDaysAndStation(fileInfoDT)
}
\arguments{
\item{fileInfoDT}{Data Table with image infromation}
}
\description{
Finds if an observation falls into the day time or not
}

