% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageStatistics.R
\name{UniqueDaysPerStation}
\alias{UniqueDaysPerStation}
\title{Finds if an observation falls into the day time or not}
\usage{
UniqueDaysPerStation(fileInfo)
}
\arguments{
\item{fileInfo}{Data Table with image infromation}
}
\description{
Finds if an observation falls into the day time or not
}

