% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRelated.R
\name{WindowFilterDayLightHours}
\alias{WindowFilterDayLightHours}
\title{Filter the lighthours given the information of the station (longitude,
latitude) and day}
\usage{
WindowFilterDayLightHours(fileInfo, properties, offsetBeforeSunrise,
  offsetAfterSunrise, offsetBeforeSunset, offsetAfterSunset)
}
\arguments{
\item{fileInfo}{data table with image files information}

\item{properties}{data table with specific information}

\item{offsetBeforeSunrise}{offset to include time before sunrise in minutes}

\item{offsetAfterSunrise}{offset to include time after sunrise in minutes}

\item{offsetBeforeSunset}{offset to include time before sunset in minutes}

\item{offsetAfterSunset}{offset to include time after sunset in minutes}
}
\description{
Filter the lighthours given the information of the station (longitude,
latitude) and day
}

