% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageStatistics.R
\name{ExtractBasicImageStatistics}
\alias{ExtractBasicImageStatistics}
\title{Extract basic image attributes}
\usage{
ExtractBasicImageStatistics(fullFilename, pattern = "na*me_yyyymmdd_hhmm.jpg")
}
\arguments{
\item{fullFilename}{Full path to the file}

\item{pattern}{(to extract name, date, and time)}
}
\value{
vector of name, date, time, mean, and variance
}
\description{
Extract basic image attributes
}

