% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dehazing.R
\name{GetAtmosphere}
\alias{GetAtmosphere}
\title{Obtains atmosphere}
\usage{
GetAtmosphere(image, darkChannel = NULL, winSize = 15)
}
\arguments{
\item{image}{The image object}

\item{darkChannel}{Image dark channel (defaults to default dark channel)}
}
\description{
Obtains atmosphere
}
