% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{GetHorizAvgTrans}
\alias{GetHorizAvgTrans}
\title{Obtains the dark horizontal average transmission}
\usage{
GetHorizAvgTrans(image, winSize, omega, lambda)
}
\arguments{
\item{image}{The image object}

\item{winSize}{Should probably be renamed}
}
\value{
vertical transmission profile
}
\description{
Obtains the dark horizontal average transmission
}

