% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portingMatlabCode.R
\name{GetRadiance}
\alias{GetRadiance}
\title{Obtains transmission}
\usage{
GetRadiance(image, transmission, atmosphere)
}
\arguments{
\item{image}{The image object}

\item{transmission}{The image transmission}

\item{atmosphere}{The image atmosphere}
}
\description{
Obtains transmission
}

