% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{data}
\name{properties}
\alias{properties}
\title{Properties of camera}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{locationId}{indentifier of camera, integer}
  \item{stationId}{identifier of camera station KNMI id, string}
  ...
}}
\usage{
properties
}
\description{
Obtained by read.csv("inst/extdata/properties.csv", stringsAsFactors = FALSE)
}
\keyword{datasets}

