% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haziness.R
\name{TransmissionChangepoint}
\alias{TransmissionChangepoint}
\title{Obtains change point of transmission}
\usage{
TransmissionChangepoint(transmission)
}
\arguments{
\item{transmission}{The horizontally averaged transmission}
}
\description{
Obtains change point of transmission
}

